Overloaded operator<< for the small integral types.
Without these instances, the char/unsigned char operator<< is used
instead, which is completely wrong.

I am surprised that the compiler doesn't complain about duplicate
overloadings with the addition of these.

+++ src/tiles/TilesState.cpp
@@ -7,10 +7,10 @@ std::ostream & operator <<(std::ostream &o, const TilesState15 &tiles)
   for (unsigned y = 0; y < 4; y += 1) {
       unsigned x = 0;
       do {
-        o << static_cast<int>(tiles(y, x)) << '\t';
+        o << tiles(y, x) << '\t';
         x += 1;
       } while (x < 3);
-      o << static_cast<int>(tiles(y, 3)) << '\n';
+      o << tiles(y, 3) << '\n';
   }
 
   return o;
